Start using goog.module (fixes #4)
Introduce necessary changes in the build steps. Application entry point
and RootView are the first modules converted. This change introduces a
very nice side effect; we just goog.require the entry point in index.html.
For bundle builds, this call is removed and it still works.

This also helps us eliminate the nasty `goog.exportSymbol` call that was
previously necessary for the bootstrapper.

Wow. I love this goog.module!