Rewrite all the classes to use goog.module (#4)
I love how the code looks. Very concise!

This commit also includes TodoModel replaced with a single instance
instead of a singleton. This is a drawback — the class is not lazy
instantiated, but hey... it's the bare-bone model anyway, the "single"
single source of truth (as in, there is no other models :) )